Add executable for the recording server
Rather than having to call "python3" and specify the module an
executable to be called just as "nextcloud-talk-recording" was added for
convenience.

However, right now it is not possible to just specify a module to be
loaded; as a function is needed to entry module is now wrapped in a
function that is explicitly called for the executable and implicitly
called when the module is loaded, thus keeping the previous behaviour.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>