Add systemd service for the recording server
The recording server does not need to be run as root, although it needs
a home directory to store some Firefox and Pulseaudio related files (not
explicitly created by the recording server but by the services
themselves). Therefore a system user, which can not log in, but with a
home directory, is used to run it.

Unfortunately the service is named "python3-nextcloud-talk-recording",
just like the package, as renaming the service requires a "debian/rules"
file with a content like:

override_dh_installsystemd:
        dh_installsystemd --name=nextcloud-talk-recording

but stdeb provides its own "debian/rules" files and there does not seem
to be a way to extend/customize it.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>